add useGame hook
